requirements.txt
django
mysqlclient
PyMuPDF 
python-docx
djangorestframework
django-cors-headers
docx2txt
scikit-learn
prometheus-client
openai